1117 API 연동 기능 마무리

+++ src/main/java/com/project/wwg/plan/dto/Spot.java
@@ -2,8 +2,6 @@ package com.project.wwg.plan.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,8 +9,8 @@ public class Spot {
     @NonNull
     private String title;
     private String info;
-    private BigDecimal lat;
-    private BigDecimal lng;
+    private double lat;
+    private double lng;
     private String address;
     private String photo;
     private String phone;
